refactor(core): promote pkg/childproc to internal/core/childproc (no behavior change)
