+++ internal/core/childproc/childproc_test.go
